pyproject.toml
[project]
name = "octoopt"
version = "0.1.0"
description = "octo automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]


[tool.ruff]
line-length = 88
